backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aioredis==2.0.1
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.5.14
ecdsa==0.19.1
fastapi==0.95.2
frozenlist==1.6.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
langchain==0.0.198
langchainplus-sdk==0.0.20
loguru==0.7.0
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numexpr==2.10.2
numpy==1.26.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==25.0
passlib==1.7.4
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==4.5.5
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.22.0
wheel==0.45.1
win32_setctime==1.2.0
yarl==1.20.0
